+++ bundle/lib/instance.dart
@@ -1,0 +1,8 @@
+/// A library manage [DeviceVendorInfo] instance.
+library instance;
+
+import 'src/instance.dart';
+
+export 'package:device_vendor_info_interface/interface.dart'
+    show DeviceVendorInfoLoader;
+export 'src/instance.dart';