+++ bundle/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+* Move `DeviceVendorInfo` class under `instance.dart` libarary.
+* Update example using direct callback function.
+* Provide export hardware information to JSON object in Dart.
+* Add topics
+
 ## 0.1.1
 
 * Add three functions to fetch hardware informations without calling instance.